requirements.txt
Flask==3.1.2
spacy==3.8.4
simstring-fast==0.6.0
pandas==3.0.1
pyodbc==5.2.0
werkzeug==3.1.3
python-dotenv==1.0.1
